refactor(auto): rename sccexporter automation to connecttelemetry
The 'sccexporter' name describes a narrower role (export-only) than the
plugin is growing into (return/control, cloud-driven config), and the
name is baked into the public 'type' config contract and the allautos
factory key -- cheapest to change now, before it is widely deployed.

Renames the package/dir pkg/auto/sccexporter -> pkg/auto/connecttelemetry,
the AutoName/type string, the factory registration, and all doc/example/
gen-dbo references. Runtime .data configs are intentionally left untouched.
