+++ pkg/auto/connecttelemetry/auto_test.go
@@ -1,4 +1,4 @@
-package sccexporter
+package connecttelemetry
 
 import (
 	"context"
